+++ media_agent/image/__init__.py
@@ -1,0 +1,3 @@
+from media_agent.image.generator import ImageGenerator
+
+__all__ = ["ImageGenerator"]